Draw a line between current and previous drag position (#68)
This implementation is suuuuperrrr...ssssslooooooowww.  Every setpixel
dispatches a redux action and it goes through the whole state
hierarchy.  But perf is surprisingly decent anyway and if it bothers
someone, it's easy to optimize by adding a DRAW_LINE action in
addition to SET_PIXEL.

This implementation does not perform a line draw between coordinates
when painting with a brush.  I just felt like this could be too slow
given how crappy my implementation is.

fixes #68