Fix out-of-bounds charcode (#95)
I can't repro this bug on macOS -- it might be resolution dependent.

The problem is that I'm treating the "isActive" flag coming from
react-cursor-position as meaning "inside/outside of div area" which is
usually the right assumption.  But it seems like due to round off (or
something else), sometimes when isActive is true, cursor positions
outside of the tracked div get reported.  Or well, maybe it's an
ordering problem (charpos gets set to state first, followed by
isActive==false).