Significantly improve canvas perf by reducing cache invalidations (#90)
CharGrid caches a bunch of things (like the font bitmaps for <canvas>)
and relies on input behaving reasonably.  The earlier commit started
triggering full cache invalidates as the input prop "font" was
allocated a new object on every setPixel call.

To fix this, simply memoize the font bits object so that the same
memory reference is returned when the inputs don't actually change.