Export multiple screens as a gif anim (#44)
This is sort of WIP -- the gif exporter is SUUUUPER slow (or I'm using
it wrong) and even crashes on larger exports with the default 64kb
internal buffer size.  So need to either optimize gif-encoder or find
a better library for this.

Also there's some kind of an indexing bug in the export code (could
affect .asm and other exports too).  It sometimes crashes with
framebuf being null.  I suspect there's a screen <-> framebuf indexing
bug somewhere.  Probably repros with reordered tabs + some deleted
screens.