Add settings option to restrict canvas/char selector window relative scaling (#18)
If set, only allow window resizing the scale the canvas and the
character selector by 1x, 1.5x, 2x, 2.5x etc.  This makes it easier to
resize the window so that the scale snaps to a full integer scale.  An
integer scale means no bitmap scaling artifacts in the canvas.

fixes #18