Fix weird rotation related char selection bugs (#103)
Biggest impact change is the identity check (see code comments).

But while debugging this, found a couple of other rotation related
bugs.  Basically the transformation needs to be applied in redux
reducers too when setting the selectedChar field.  This whole rotation
business was maybe designed in a wrong way.

fixes #103