TypeScriptify selectors and the CharSelect component
This was trickiest to type so far.  The problem here is that any
actions touching frame buffers get routed in the framebufList reducer
based on the framebufIndex field.  This framebufIndex field is plugged
into actions implicitly as part of connect() call in the merge props
function.

The action creators all have framebufIndex in their type but when we
call these creators from a component, we don't have the framebufIndex
as it's inserted in merge props.  So this patch contains a bunch of
magic to convert the action creator types to something that can be
used in connected React components.