Get rid of the horrible withMouseCharPositionShiftLockAxis abomination
This was a really hairy up/down props passing, state messing, callback
kind of a HOC which didn't really make much sense and was used only by
one component.  So move the code into the FrameBufferView component
and just accept it that the logic is a little complicated.

Gets rid of react-cursor-position dependency which I never quite liked
either.  The code to do this is trivial but having it as a component
means we add extra HOCs which again make it hard to track what's going
on.