Enable app shortcuts when it gains focus
Previous code tried to do that too, but this was not working on Linux.

And the code indeed looks a little wonky, it was handling browser-window-blur
but not browser-window-focus.  I think when I worked on this
on macOS, this code was alright.  But looks like on Linux, there's only
one browser-window-blur and a browser-window-focus after closing a dialog, not
a separate 'focus' event.

for #161